feat(ContributorsTable): upgrade ContributorsList to ContributorsTable
